feat: consolidate core config and routing modules
