+++ docs/gate-review-comment-contract.md
@@ -30,7 +30,7 @@ This contract covers exactly two gates with distinct lifecycle semantics:
 
 These gates are related but **not interchangeable**.
 
-Each gate's review angles are defined in the project config (`gates.draft.angles` and `gates.preApproval.angles` in `.pi/dev-loop/defaults.yaml`). The reviewer persona for each angle is resolved via `resolveReviewerRole` from the persona registry (`packages/core/src/config/roles.mjs`). Consumer repos may override angles and map custom personas via their own config.
+Each gate's review angles are defined in the project config (`gates.draft.angles` and `gates.preApproval.angles` in `.pi/dev-loop/defaults.yaml`). The reviewer persona for each angle is resolved via `resolveReviewerRole` from the persona registry (`packages/core/src/config/config.mjs`). Consumer repos may override angles and map custom personas via their own config.
 
 Resolve angles at runtime with `resolveGateAngles(config, "draft")` and `resolveGateAngles(config, "preApproval")` from `@pi-dev-loops/core/config`. Do not hardcode angle names in skill procedures or review prompts.
