+++ packages/core/package.json
@@ -5,10 +5,7 @@
   "description": "Shared deterministic support package for Pi dev-loop skills, repo-local scripts, and GitHub automation.",
   "exports": {
     "./bash-exit-one": "./src/bash-exit-one.mjs",
-    "./config": "./src/config/index.mjs",
-    "./config/schema": "./src/config/schema.mjs",
-    "./config/loader": "./src/config/loader.mjs",
-    "./config/roles": "./src/config/roles.mjs",
+    "./config": "./src/config/config.mjs",
     "./github/copilot-helpers": "./src/github/copilot-helpers.mjs",
     "./github/repo-slug": "./src/github/repo-slug.mjs",
     "./github/review-threads": "./src/github/review-threads.mjs",
@@ -17,11 +14,9 @@
     "./loop/copilot-ci-status": "./src/loop/copilot-ci-status.mjs",
     "./loop/copilot-loop-iterations": "./src/loop/copilot-loop-iterations.mjs",
     "./loop/copilot-loop-state": "./src/loop/copilot-loop-state.mjs",
-    "./loop/outer-loop-state": "./src/loop/outer-loop-state.mjs",
     "./loop/phase-files": "./src/loop/phase-files.mjs",
     "./loop/pr-gate-coordination": "./src/loop/pr-gate-coordination.mjs",
     "./loop/public-dev-loop-routing": "./src/loop/public-dev-loop-routing.mjs",
-    "./loop/retrospective-checkpoint": "./src/loop/retrospective-checkpoint.mjs",
     "./loop/reviewer-loop-state": "./src/loop/reviewer-loop-state.mjs",
     "./loop/run-inspection": "./src/loop/run-inspection.mjs",
     "./loop/steering": "./src/loop/steering.mjs",
