--- packages/core/src/config/loader.mjs
@@ -1,246 +1,0 @@
-import { readFile } from "node:fs/promises";
-import path from "node:path";
-import { parse as parseYaml } from "yaml";
-import { BUILT_IN_DEFAULTS, DevLoopConfigSchema, FileConfigSchema } from "./schema.mjs";
-
-// ============================================================================
-// Error types
-// ============================================================================
-
-/**
- * @typedef {object} ConfigLoadError
- * @property {string} path - Human-readable file path or layer name
- * @property {string} message - Error description
- * @property {"defaults"|"settings"|"merged"} layer - Which config layer failed
- */
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-/**
- * Merge two config objects. Keys in `source` override keys in `target`.
- * Family objects merge at one level, except `gates`, which merges one extra
- * nested gate-object level so settings can override `draft.requireCi` without
- * restating the shipped draft angles.
- * @param {Record<string, unknown>} target
- * @param {Record<string, unknown>} source
- * @returns {Record<string, unknown>}
- */
-function mergeConfigLayers(target, source) {
-  const result = { ...target };
-  for (const key of Object.keys(source)) {
-    if (
-      key !== "version" &&
-      typeof source[key] === "object" &&
-      source[key] !== null &&
-      !Array.isArray(source[key]) &&
-      typeof result[key] === "object" &&
-      result[key] !== null &&
-      !Array.isArray(result[key])
-    ) {
-      result[key] = key === "gates"
-        ? mergeNestedObject(result[key], source[key])
-        : { ...(result[key] || {}), ...(source[key] || {}) };
-    } else {
-      result[key] = source[key];
-    }
-  }
-  return result;
-}
-
-function mergeNestedObject(target, source) {
-  const result = { ...(target || {}) };
-
-  for (const key of Object.keys(source || {})) {
-    if (
-      typeof source[key] === "object" &&
-      source[key] !== null &&
-      !Array.isArray(source[key]) &&
-      typeof result[key] === "object" &&
-      result[key] !== null &&
-      !Array.isArray(result[key])
-    ) {
-      result[key] = { ...(result[key] || {}), ...(source[key] || {}) };
-    } else {
-      result[key] = source[key];
-    }
-  }
-
-  return result;
-}
-
-/**
- * Try to read and parse a config file (YAML preferred, JSON fallback).
- * Detects format from file extension: .yaml/.yml → YAML, .json → JSON.
- * Returns the parsed object or null if the file doesn't exist.
- * Throws on read errors other than ENOENT.
- * @param {string} filePath
- * @returns {Promise<object|null>}
- */
-async function readConfigFile(filePath) {
-  let raw;
-  try {
-    raw = await readFile(filePath, "utf8");
-  } catch (err) {
-    if (err.code === "ENOENT") return null;
-    throw configError(`Cannot read config file: ${err.message}`, err.code, filePath);
-  }
-
-  if (raw.trim() === "") {
-    throw configError("Config file is empty", "EMPTY_FILE", filePath);
-  }
-
-  const isYaml = filePath.endsWith(".yaml") || filePath.endsWith(".yml");
-  let parsed;
-  try {
-    parsed = isYaml ? parseYaml(raw) : JSON.parse(raw);
-  } catch (err) {
-    const format = isYaml ? "YAML" : "JSON";
-    throw configError(`Invalid ${format} in config file: ${err.message}`, `INVALID_${format.toUpperCase()}`, filePath);
-  }
-
-  if (parsed === null || typeof parsed !== "object" || Array.isArray(parsed)) {
-    throw configError("Config file must be an object", "NOT_AN_OBJECT", filePath);
-  }
-
-  return parsed;
-}
-
-/**
- * Find a config file by trying one or more base names in order.
- * Each base name prefers YAML (.yaml, then .yml) before JSON.
- * @param {string|string[]} basePaths - Path(s) without extension (e.g. .../defaults)
- * @returns {Promise<{ path: string, data: object|null }>}
- */
-async function findConfigFile(basePaths) {
-  const candidates = Array.isArray(basePaths) ? basePaths : [basePaths];
-
-  for (const basePath of candidates) {
-    for (const ext of [".yaml", ".yml", ".json"]) {
-      const filePath = basePath + ext;
-      const data = await readConfigFile(filePath);
-      if (data !== null) return { path: filePath, data };
-    }
-  }
-
-  return { path: candidates[0] + ".yaml", data: null };
-}
-
-/**
- * @param {string} message
- * @param {string} code
- * @param {string} filePath
- * @returns {Error & { code: string, path: string }}
- */
-function configError(message, code, filePath) {
-  return Object.assign(new Error(message), { code, path: filePath });
-}
-
-/**
- * Try to load and merge one config layer (defaults or settings).
- * @param {Record<string, unknown>} merged - Current merged config
- * @param {string|string[]} basePaths - Config file base path(s) without extension
- * @param {"defaults"|"settings"} layer - Layer name
- * @param {string[]} warnings
- * @param {ConfigLoadError[]} errors
- * @param {{ warnOnMissing?: boolean }} [options]
- * @returns {Promise<Record<string, unknown>>}
- */
-async function applyLayer(merged, basePaths, layer, warnings, errors, options = {}) {
-  let filePath, data = null;
-  try {
-    const found = await findConfigFile(basePaths);
-    filePath = found.path;
-    data = found.data;
-  } catch (err) {
-    const preferredBasePath = Array.isArray(basePaths) ? basePaths[0] : basePaths;
-    const errorPath = err.path ?? preferredBasePath + ".yaml";
-    errors.push({
-      path: errorPath,
-      message: `${path.basename(errorPath)}: ${err.message}`,
-      layer,
-    });
-    return merged;
-  }
-
-  if (data === null) {
-    if (options.warnOnMissing) {
-      warnings.push(`Committed ${layer} config not found (tried .yaml, .yml, and .json), using built-in defaults`);
-    }
-    return merged;
-  }
-
-  // Validate the file's structure before merging
-  const validation = FileConfigSchema.safeParse(data);
-  if (!validation.success) {
-    errors.push({
-      path: filePath,
-      message: `${path.basename(filePath)}: Schema validation failed: ${validation.error.issues.map(i => `${i.path.join(".")}: ${i.message}`).join("; ")}`,
-      layer,
-    });
-    return merged;
-  }
-
-  return mergeConfigLayers(merged, data);
-}
-
-// ============================================================================
-// Loader
-// ============================================================================
-
-/**
- * @typedef {object} LoadResult
- * @property {import("./schema.mjs").DevLoopConfig} config
- * @property {string[]} warnings
- * @property {ConfigLoadError[]} errors
- */
-
-/**
- * @typedef {object} LoadOptions
- * @property {string} [repoRoot] - Path to repository root (default: process.cwd())
- */
-
-/**
- * Load the dev-loop configuration with full precedence:
- *   settings.(yaml|yml|json) > legacy overrides.(yaml|yml|json) > defaults.(yaml|yml|json) > built-in defaults
- *
- * Never throws for config-related problems.
- * Returns built-in defaults even when all files are missing or broken.
- *
- * @param {LoadOptions} [options]
- * @returns {Promise<LoadResult>}
- */
-export async function loadDevLoopConfig(options = {}) {
-  const repoRoot = options.repoRoot ?? process.cwd();
-  const configDir = path.join(repoRoot, ".pi", "dev-loop");
-  const defaultsPath = path.join(configDir, "defaults");
-  const settingsPaths = [path.join(configDir, "settings"), path.join(configDir, "overrides")];
-
-  /** @type {string[]} */
-  const warnings = [];
-  /** @type {ConfigLoadError[]} */
-  const errors = [];
-
-  let merged = { ...BUILT_IN_DEFAULTS };
-
-  merged = await applyLayer(merged, defaultsPath, "defaults", warnings, errors, {
-    warnOnMissing: true,
-  });
-
-  merged = await applyLayer(merged, settingsPaths, "settings", warnings, errors);
-
-  // Validate final merged config
-  const result = DevLoopConfigSchema.safeParse(merged);
-  if (!result.success) {
-    errors.push({
-      path: "<merged>",
-      message: `Config validation failed: ${result.error.issues.map(i => `${i.path.join(".")}: ${i.message}`).join("; ")}`,
-      layer: "merged",
-    });
-    // Return merged as-is — caller gets validation errors but still has config with all layers applied
-    return { config: /** @type {*} */ (merged), warnings, errors };
-  }
-
-  return { config: result.data, warnings, errors };
-}
