+++ scripts/loop/print-gates.mjs
@@ -20,9 +20,9 @@ import process from "node:process";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 
-import { loadDevLoopConfig } from "../../packages/core/src/config/loader.mjs";
-import { resolveGateConfig } from "../../packages/core/src/config/model-resolution.mjs";
-import { resolveReviewerRole } from "../../packages/core/src/config/roles.mjs";
+import { loadDevLoopConfig } from "../../packages/core/src/config/config.mjs";
+import { resolveGateConfig } from "../../packages/core/src/config/config.mjs";
+import { resolveReviewerRole } from "../../packages/core/src/config/config.mjs";
 
 async function run({ stdout = process.stdout, repoRoot = process.cwd() } = {}) {
   const { config } = await loadDevLoopConfig({ repoRoot });
